python-levenstein replaced sequencematcher

+++ analysis/paper/ukb_efo.py
@@ -12,6 +12,8 @@ import timeit
 import scispacy
 import spacy
 import difflib
+from fuzzywuzzy import fuzz
+import Levenshtein
 from scripts.vectology_functions import create_aaa_distances, create_pair_distances, embed_text, encode_traits, create_efo_nxo, create_efo_data
 from loguru import logger
 from pathlib import Path
@@ -281,6 +283,7 @@ def run_seq_matcher(ebi_df,efo_node_df):
         logger.info(f'\n{efo_node_df.head()}')
         ebi_list = list(ebi_df['query'])[:10]
         efo_list = list(efo_node_df['efo_label'])[:10]
+        logger.info(ebi_list)
         #for i in range(0,len(ebi_dic)):
         #    if i % 100 == 0:
         #        logger.info(i)
@@ -291,6 +294,26 @@ def run_seq_matcher(ebi_df,efo_node_df):
         logger.info(distance)
                 #distance = create_edit_distance(query,efo_label)
         #d.append({'mapping_id':ebi_dic[i]['mapping_id'],'manual':ebi_dic[i]['full_id'],'prediction':efo_dic[j]['efo_id'],'score':distance})
+        df = pd.DataFrame(d)
+        logger.info(df.head())
+        df.to_csv(f,sep='\t',index=False,compression='gzip')
+
+def run_levenshtein(ebi_df,efo_node_df):
+    f = f'{output}/levenshtein-pairwise.tsv.gz'
+    if os.path.exists(f):
+        logger.info(f'{f} done')
+    else:
+        d = []
+        ebi_dic = ebi_df.to_dict('records')
+        efo_dic = efo_node_df.to_dict('records') 
+        for i in range(0,len(ebi_dic)):
+            if i % 100 == 0:
+                logger.info(i)
+            for j in range(0,len(efo_dic)):
+                query = ebi_dic[i]['query']
+                efo_label = efo_dic[j]['efo_label']
+                distance =  Levenshtein.ratio(query, efo_label)
+                d.append({'mapping_id':ebi_dic[i]['mapping_id'],'manual':ebi_dic[i]['full_id'],'prediction':efo_dic[j]['efo_id'],'score':distance})
         df = pd.DataFrame(d)
         logger.info(df.head())
         df.to_csv(f,sep='\t',index=False,compression='gzip')
@@ -754,7 +777,8 @@ def dev():
     efo_node_df = efo_node_data_v1()
     # get ebi efo data
     ebi_df = get_ebi_data(efo_node_df)
-    run_seq_matcher(ebi_df,efo_node_df)
+    #run_seq_matcher(ebi_df,efo_node_df)
+    run_levenshtein(ebi_df,efo_node_df)
 
 if __name__ == "__main__":
     #run()
